Added a new test. Refactored code.| #2471
